refactor: improve auth using Pages router instead of App router
