--- frontend/src/pages/layout.tsx
@@ -1,32 +1,0 @@
-import type { Metadata } from "next"
-import { Poppins } from "next/font/google"
-import "./globals.css"
-import SessionWrapper from "@components/SessionWrapper"
-
-const poppins = Poppins({
-  subsets: ["latin"],
-  weight: ["400", "500", "600", "700"],
-})
-
-export const metadata: Metadata = {
-  title: "Festwrap",
-  description: "Spotify playlist generator for music festivals",
-}
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
-  return (
-    <SessionWrapper>
-      <html lang="en">
-        <body className={poppins.className}>
-          <div className="flex flex-col h-full items-center justify-center bg-gray-50">
-            {children}
-          </div>
-        </body>
-      </html>
-    </SessionWrapper>
-  )
-}
