+++ frontend/src/styles/globals.css
@@ -3,6 +3,6 @@
 @tailwind utilities;
 
 html,
-body {
+body, #__next {
   height: 100%;
 }
