+++ frontend/vitest.config.mts
@@ -11,7 +11,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
-      "@components": path.resolve(__dirname, "./src/app/components"),
+      "@components": path.resolve(__dirname, "./src/components"),
     },
   },
 })
